+++ integration-tests/src/test/java/org/apache/druid/tests/indexer/ITS3ToHdfsParallelIndexTest.java
@@ -33,18 +33,19 @@ import java.util.List;
  * 1) Set the bucket and path for your data. This can be done by setting -Ddruid.test.config.cloudBucket and
  *    -Ddruid.test.config.cloudPath or setting "cloud_bucket" and "cloud_path" in the config file.
  * 2) Copy wikipedia_index_data1.json, wikipedia_index_data2.json, and wikipedia_index_data3.json
- *    located in integration-tests/src/test/resources/data/batch_index to your GCS at the location set in step 1.
- * 3) Provide -Doverride.config.path=<PATH_TO_FILE> with gcs configs set. See
- *    integration-tests/docker/environment-configs/override-examples/gcs for env vars to provide.
- * 4) Provide -Dresource.file.dir.path=<PATH_TO_FOLDER> with folder that contains GOOGLE_APPLICATION_CREDENTIALS file
+ *    located in integration-tests/src/test/resources/data/batch_index to your S3 at the location set in step 1.
+ * 3) Provide -Doverride.config.path=<PATH_TO_FILE> with s3 credentials and hdfs deep storage configs set. See
+ *    integration-tests/docker/environment-configs/override-examples/s3 and
+ *    integration-tests/docker/environment-configs/override-examples/hdfs for env vars to provide.
+ * 4) Run the test with -Dstart.hadoop.docker=true in the mvn command
  */
-@Test(groups = TestNGGroup.GCS_DEEP_STORAGE)
+@Test(groups = TestNGGroup.HDFS_DEEP_STORAGE)
 @Guice(moduleFactory = DruidTestModuleFactory.class)
-public class ITGcsToGcsParallelIndexTest extends AbstractGcsInputSourceSimpleIndexTest
+public class ITS3ToHdfsParallelIndexTest extends AbstractS3InputSourceSimpleIndexTest
 {
   @Test(dataProvider = "resources")
-  public void testGcsIndexData(Pair<String, List> gcsInputSource) throws Exception
+  public void testS3IndexData(Pair<String, List> s3InputSource) throws Exception
   {
-    doTest(gcsInputSource);
+    doTest(s3InputSource);
   }
 }
